# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Nada UTP TIS",
        "version": "1.0.0"
    },
    "paths": {
        "/api/test": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Test endpoint",
                "operationId": "af4a902acfa310be105fc0fa630376ca",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get all items",
                "operationId": "2cbf07eb95d033f8775c36424ac8fa06",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Create item",
                "operationId": "af37857663e6805dd367a55f374c9ee1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "harga"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get item by ID",
                "operationId": "7f1ffbd9fa2b704c61ebcccb636ebb27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Update full item",
                "operationId": "4ac9fb55d8a18bd1cba3a29d84f9fd4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "harga"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete item",
                "operationId": "0f88b60b73861d5414f8ea2fedd3dfc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Items"
                ],
                "summary": "Update partial item",
                "operationId": "8171e1e0bd396fd89495538eb21a8869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "harga": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Test",
            "description": "Test"
        },
        {
            "name": "Items",
            "description": "Items"
        }
    ]
}